fix: support cluster mode
When running in cluster mode, redis reports the following error:
CROSSSLOT Keys in request don't hash to the same slot

The evalsha line treats rate, burst, now, and duration as KEYS, which
forces redis to hash them and determine they would hash to separate
slots.

However, we are never using these parameters as keys, so we can safely
convert this call to using 1 key and 4 arguments.